Use ReadmeAnyFromPod for README in build directory
ReadmeFromPod takes the POD text from the main ".pm" file even when a ".pod" file alternative exists. ReadmeAnyFromPod does not suffer from this problem, and it is already used in Milla to generate the Markdown file in the root directory (so it's nothing new).

This proposal changes ReadmeFromPod to ReadmeAnyFromPod, with the right plugin name "ReadmeAnyFromPod/ReadmeTextInBuild" that creates a "README" file inside the build directory.